Parameter error on server startup.

+++ src/main/java/Controllers/CSController.java
@@ -1,34 +1,87 @@
 package Controllers;
 import Server.Main;
+import com.sun.jersey.multipart.FormDataParam;
+import org.json.simple.JSONArray;
+import org.json.simple.JSONObject;
+
+import javax.ws.rs.*;
+import javax.ws.rs.core.MediaType;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 
 public class CSController {
     //Method which lists the computer science
-    public static void ListCS(){
+    @GET
+    @Path("list")
+    @Produces(MediaType.APPLICATION_JSON)
+    public String ListCS(){
+        System.out.println("cs/list");
+        JSONArray list = new JSONArray();
         try{
             PreparedStatement ps = Main.db.prepareStatement("SELECT QuestionID, Subtopic, Question, AnswerA, AnswerB, AnswerC, AnswerD FROM CS");
             ResultSet results = ps.executeQuery();
             while(results.next()){    //Method keeps getting data until it reaches the end of the database.
-                int QuestionID = results.getInt(1);
-                String Subtopic = results.getString(2);
-                String Question = results.getString(3);
-                String AnswerA = results.getString(4);
-                String AnswerB = results.getString(5);
-                String AnswerC = results.getString(6);
-                String AnswerD = results.getString(7);
-                System.out.println(QuestionID + " " + Subtopic + " " + Question + " " + AnswerA + " " + AnswerB + " " + AnswerC + " " + AnswerD);
+                JSONObject item = new JSONObject();
+                item.put("QuestionID", results.getInt(1));
+                item.put("Subtopic", results.getString(2));
+                item.put("Question", results.getString(3));
+                item.put("AnswerA", results.getString(4));
+                item.put("AnswerB", results.getString(5));
+                item.put("AnswerC", results.getString(6));
+                item.put("AnswerD", results.getString(7));
 
             }
+            return list.toString();
         }catch (Exception exception){
             System.out.println("Database error: " + exception.getMessage());
+            return "{\"error\": \"Unable to list items, please see server console for more info.\"}";
+        }
+    }
+
+    @GET
+    @Path("get/{QuestionID}")
+    @Produces(MediaType.APPLICATION_JSON)
+    public String getThing(@PathParam("QuestionID") Integer QuestionID) throws Exception {
+        if(QuestionID==null){
+            throw new Exception("Thing's 'id' is missing in the HTTP request's URL.");
+
+        }
+        System.out.println("cs/get/" + QuestionID);
+        JSONObject item = new JSONObject();
+        try{
+            PreparedStatement ps = Main.db.prepareStatement("Select Subtopic, Question, AnswerA, AnswerB, AnswerC, AnswerD FROM CS WHERE QuestionID =?");
+            ps.setInt(1,QuestionID);
+            ResultSet results = ps.executeQuery();
+            if (results.next()){
+                item.put("QuestionID", QuestionID);
+                item.put("Subtopic", results.getString(1));
+                item.put("Question",results.getString(2));
+                item.put("AnswerA", results.getString(3));
+                item.put("AnswerB", results.getString(4));
+                item.put("AnswerC", results.getString(5));
+                item.put("AnswerD", results.getString(6));
+            }
+            return item.toString();
+
+        }catch (Exception exception){
+            System.out.println("Database error: " + exception.getMessage());
+            return "{\"error\": \"Unable to get item, please see server console for more info.\"}";
         }
     }
 
     //Method which inserts a record into the computer science table
-    public static void InsertIntoCS(int QuestionID, String Question, String Subtopic, String AnswerA, String AnswerB, String AnswerC, String AnswerD){
+    @POST
+    @Path("new")
+    @Consumes(MediaType.MULTIPART_FORM_DATA)
+    @Produces(MediaType.APPLICATION_JSON)
+    public String InsertIntoCS(@FormDataParam("QuestionID") Integer QuestionID, @FormDataParam("Subtopic") String Subtopic, @FormDataParam("Question") String Question, @FormDataParam("AnswerA") String AnswerA, @FormDataParam("AnswerB") String AnswerB, @FormDataParam("AnswerC") String AnswerC, @FormDataParam("AnswerD") String AnswerD){
         try{
+            if (QuestionID == null || Subtopic == null || Question == null || AnswerA == null || AnswerB == null || AnswerC ==null || AnswerD == null){
+                throw new Exception("One or more form data parameters are missing in the HTTP request.");
+            }
+            System.out.println("cs/newid=" + QuestionID);
             PreparedStatement ps = Main.db.prepareStatement("INSERT INTO CS(QuestionID, Subtopic, Question, AnswerA, AnswerB, AnswerC, AnswerD) VALUES (?,?,?,?,?,?,?)");
+
             ps.setInt(1, QuestionID);
             ps.setString(2, Subtopic);
             ps.setString(3, Question);
@@ -36,43 +89,64 @@ public class CSController {
             ps.setString(5, AnswerB);
             ps.setString(6, AnswerC);
             ps.setString(7, AnswerD);
-            ps.executeUpdate();
-            System.out.println("Record added to Computer Science Table");
+            ps.execute();
+            return "{\"status\": \"OK\"}";
 
         }catch (Exception exception){
-            System.out.println(exception.getMessage());
-            System.out.println("Error: Something has gone wrong");
+            System.out.println("Database error: " + exception.getMessage());
+            return "{\"error\": \"Unable to create new item, please see server console for more info.\"}";
         }
     }
 
     //Updates a record in the computer science table
-    public static void UpdateCS (String Subtopic, String Question, String AnswerA, String AnswerB, String AnswerC, String AnswerD){
+    @POST
+    @Path("update")
+    @Consumes(MediaType.MULTIPART_FORM_DATA)
+    @Produces(MediaType.APPLICATION_JSON)
+    public String UpdateCS(@FormDataParam("QuestionID") Integer QuestionID, @FormDataParam("Subtopic") String Subtopic, @FormDataParam("Question") String Question, @FormDataParam("AnswerA") String AnswerA,@FormDataParam("AnswerB") String AnswerB,@FormDataParam("AnswerC") String AnswerC, @FormDataParam("AnswerD") String AnswerD){
         try{
-            PreparedStatement ps = Main.db.prepareStatement("UPDATE CS SET Subtopic = ?, Question = ?, AnswerA = ?, AnswerB = ?, AnswerC = ?, AnswerD = ? WHERE UserID = ?");
+            if (QuestionID == null || Subtopic == null || Question == null || AnswerA ==null || AnswerB == null || AnswerC == null || AnswerD == null){
+                throw new Exception("One or more form data parameters are missing in the HTTP request.");
+            }
+            System.out.println("cs/update QuestionID=" + QuestionID);
+            PreparedStatement ps = Main.db.prepareStatement("UPDATE CS SET Subtopic = ?, Question = ?, AnswerA = ?, AnswerB = ?, AnswerC = ?, AnswerD = ? WHERE QuestionID = ?");
             ps.setString(1,Subtopic);
             ps.setString(2,Question);
             ps.setString(3,AnswerA);
             ps.setString(4,AnswerB);
             ps.setString(5,AnswerC);
             ps.setString(6,AnswerD);
-            ps.executeUpdate();
+            ps.setInt(7,QuestionID);
+            ps.execute();
+            return "{\"status\": \"OK\"}";
 
         }catch(Exception exception){
-            System.out.println(exception.getMessage());
+            System.out.println("Database error: " + exception.getMessage());
+            return "{\"error\": \"Unable to update item please see server console for more info.\"}";
 
         }
     }
 
     //Deletes a record in the computer science table
-    public static void DeleteCS(int QuestionID){
+    @POST
+    @Path("delete")
+    @Consumes(MediaType.MULTIPART_FORM_DATA)
+    @Produces(MediaType.APPLICATION_JSON)
+    public String DeleteCS(@FormDataParam("QuestionID") Integer QuestionID){
         try{
+            if(QuestionID==null){
+                throw new Exception("One or more form data parameters are missing in the HTTP request.");
+            }
+            System.out.println("cs/delete QuestionID=" + QuestionID);
             PreparedStatement ps = Main.db.prepareStatement("DELETE FROM CS WHERE QuestionID = ?");
             ps.setInt(1,QuestionID);
-            ps.executeUpdate();
+            ps.execute();
+            return"{\"status\": \"OK\"}";
 
 
         }catch (Exception exception){
-            System.out.println(exception.getMessage());
+            System.out.println("Database error:" + exception.getMessage());
+            return "{\"error\": \"Unable to delete item, please see server console for more info.\"}";
         }
     }
